Move honeycomb singleton out of prelude
