+++ templates/boltzmann/core/middleware.ts
@@ -1,7 +1,6 @@
 void `{% if selftest %}`;
 export { Handler, Adaptor, Middleware, MiddlewareConfig, Response, buildMiddleware, handler }
-import { honeycomb } from '../core/prelude'
-import { beeline, getOtelMockSpans, otel } from '../core/honeycomb'
+import { beeline, getOtelMockSpans, honeycomb, otel } from '../core/honeycomb'
 import { HttpMetadata } from '../core/prelude'
 import { HTTPMethod } from 'find-my-way'
 import isDev from 'are-we-dev'
